// app/types/projects.ts
import { RichTextContent } from "@graphcms/rich-text-types";

export type knownTechsh = {
  iconSvg: string;
  name: string;
  startDate: string;
};

export type ProjectSection = {
  title: string;
  image: {
    url: string;
  };
};
export type Project = {
  slug: string;
  thumbnail: {
    url: string;
  };
  title: string;
  shortDescription: string;
  technology: knownTechsh[];
  pageThumbnail: {
    url: string;
  };
  sections: ProjectSection[];
  description: {
    raw: RichTextContent;
    text: string;
  };
  liveProjectUrl?: string;
  githubUrl?: string;
};